backend/requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
pydantic
python-dotenv
requests

